requirements.txt
python-telegram-bot==13.7
instaloader==4.10.2
python-dotenv==1.0.0
urllib3==1.26.15
six>=1.10
requests>=2.31.0 